Nbody works

+++ DX12_Project/DX12_Project/src/graphics/Shader.cpp
@@ -62,7 +62,7 @@ namespace dx
 		}
 
 		D3D12_INPUT_LAYOUT_DESC inputLayoutDesc = {};
-		inputLayoutDesc.NumElements = inputElementDesc.size();
+		inputLayoutDesc.NumElements = (UINT)inputElementDesc.size();
 		inputLayoutDesc.pInputElementDescs = &inputElementDesc[0];
 
 		//Fill in the pipeline description
